+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "AramianVasak",
   "version": "1.0.0",
-  "description": "",
+  "description": "https://js-av.herokuapp.com/",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -17,5 +18,10 @@
   "bugs": {
     "url": "https://github.com/Geekbrains-Frontend-Level-2/AramianVasak/issues"
   },
-  "homepage": "https://github.com/Geekbrains-Frontend-Level-2/AramianVasak#readme"
+  "homepage": "https://github.com/Geekbrains-Frontend-Level-2/AramianVasak#readme",
+  "devDependencies": {
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11"
+  },
+  "dependencies": {}
 }
