+++ public/index.html
@@ -8,7 +8,19 @@
       href="https://fonts.googleapis.com/css?family=Roboto&display=swap"
     />
     <link rel="stylesheet" href="style/style.css" />
-    <script src="js/script.js" defer></script>
   </head>
-  <body>Student was a lazy dude.</body>
+  <body>
+    <header class="header">
+      <h1>Cookies shop</h1>
+    </header>
+    <main class="main"></main>
+
+    <script src="js/Renderer.js"></script>
+    <script src="js/Item.js"></script>
+    <script src="js/Cart.js"></script>
+    <script src="js/CartItem.js"></script>
+    <script src="js/CartBtn.js"></script>
+    <script src="js/ItemsList.js"></script>
+    <script src="js/index.js"></script>
+  </body>
 </html>
