+++ public/js/index.js
@@ -1,0 +1,8 @@
+// import HEADER from "./Renderer";
+// import ItemsList from "./ItemsList";
+// import Cart from "./Cart";
+// import CartBtn from "./CartBtn";
+
+const arr = new ItemsList();
+const cart = new Cart(HEADER);
+const cartBtn = new CartBtn(HEADER, document.querySelector(".cart"));
